+++ lib/icarus/mod/tools.rb
@@ -1,12 +1,13 @@
 # frozen_string_literal: true
 
-require "firestore"
-require "github"
+require "config"
 
 module Icarus
   module Mod
     module Tools
       class Error < StandardError; end
+
+      # Icarus::Mod::Config.read
     end
   end
 end